+++ src/Ops.DataProvider.SqlServer.Promenade/Migrations/20211029220453_prom_v.0.0.152.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
+{
+    public partial class prom_v00152 : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Stub",
+                table: "Emedia");
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "Stub",
+                table: "Emedia",
+                type: "nvarchar(255)",
+                maxLength: 255,
+                nullable: false,
+                defaultValue: "");
+        }
+    }
+}