+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Podcast functionality to promenade
 - Add layout pages
 - Application permission handling and administration
+- Emedia management
 
 ### Changed
 - Default to only logging Warnings for Microsoft and System namespaces